+++ lua/plugins/lualine.lua
@@ -37,7 +37,7 @@ return {
 					},
 					{ "filetype", separator = "│" },
 					{
-						function(fmt) return require("nikero.lualine"):harpoon_widget(fmt.default_hl) end,
+						function(fmt) return require("nikero.statusline"):harpoon_widget(fmt.default_hl) end,
 						condition = function() return #require("harpoon"):list().items ~= 0 end,
 						separator = "│",
 						padding = 0,
